+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 -   Changed the repository's file structure a little, it wouldn't affect the
     runtime of the app that much
 -   Added the urls used to list down minecraft, forge, and fabric versions
+-   Added tests as I was still lazy to add and test the new stuff to the gui
 
 # Reminders (to me)